+++ .gitattributes
@@ -1,0 +1,2 @@
+# 자동 줄바꿈 처리 (OS별 차이 해결)
+* text=auto
