+++ src/screens/homescreen.js
@@ -1,0 +1,24 @@
+// src/screens/HomeScreen.js
+import React from 'react';
+import { View, Text, StyleSheet } from 'react-native';
+
+export default function HomeScreen() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>🍜 FORK 홈 화면입니다</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#fefefe',
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: '600',
+  },
+});
